Factor out network
Part of #1112
