+++ nano/node/node.hpp
@@ -6,6 +6,7 @@
 #include <nano/node/bootstrap.hpp>
 #include <nano/node/confirmation_height_processor.hpp>
 #include <nano/node/logging.hpp>
+#include <nano/node/network.hpp>
 #include <nano/node/node_observers.hpp>
 #include <nano/node/nodeconfig.hpp>
 #include <nano/node/payment_observer_processor.hpp>
@@ -13,8 +14,6 @@
 #include <nano/node/repcrawler.hpp>
 #include <nano/node/signatures.hpp>
 #include <nano/node/stats.hpp>
-#include <nano/node/transport/tcp.hpp>
-#include <nano/node/transport/udp.hpp>
 #include <nano/node/wallet.hpp>
 #include <nano/node/websocket.hpp>
 #include <nano/secure/ledger.hpp>
@@ -184,126 +183,6 @@ private:
 
 std::unique_ptr<seq_con_info_component> collect_seq_con_info (online_reps & online_reps, const std::string & name);
 
-class message_buffer final
-{
-public:
-	uint8_t * buffer{ nullptr };
-	size_t size{ 0 };
-	nano::endpoint endpoint;
-};
-/**
-  * A circular buffer for servicing nano realtime messages.
-  * This container follows a producer/consumer model where the operating system is producing data in to
-  * buffers which are serviced by internal threads.
-  * If buffers are not serviced fast enough they're internally dropped.
-  * This container has a maximum space to hold N buffers of M size and will allocate them in round-robin order.
-  * All public methods are thread-safe
-*/
-class message_buffer_manager final
-{
-public:
-	// Stats - Statistics
-	// Size - Size of each individual buffer
-	// Count - Number of buffers to allocate
-	message_buffer_manager (nano::stat & stats, size_t, size_t);
-	// Return a buffer where message data can be put
-	// Method will attempt to return the first free buffer
-	// If there are no free buffers, an unserviced buffer will be dequeued and returned
-	// Function will block if there are no free or unserviced buffers
-	// Return nullptr if the container has stopped
-	nano::message_buffer * allocate ();
-	// Queue a buffer that has been filled with message data and notify servicing threads
-	void enqueue (nano::message_buffer *);
-	// Return a buffer that has been filled with message data
-	// Function will block until a buffer has been added
-	// Return nullptr if the container has stopped
-	nano::message_buffer * dequeue ();
-	// Return a buffer to the freelist after is has been serviced
-	void release (nano::message_buffer *);
-	// Stop container and notify waiting threads
-	void stop ();
-
-private:
-	nano::stat & stats;
-	std::mutex mutex;
-	std::condition_variable condition;
-	boost::circular_buffer<nano::message_buffer *> free;
-	boost::circular_buffer<nano::message_buffer *> full;
-	std::vector<uint8_t> slab;
-	std::vector<nano::message_buffer> entries;
-	bool stopped;
-};
-class network final
-{
-public:
-	network (nano::node &, uint16_t);
-	~network ();
-	void start ();
-	void stop ();
-	void flood_message (nano::message const &);
-	void flood_vote (std::shared_ptr<nano::vote> vote_a)
-	{
-		nano::confirm_ack message (vote_a);
-		flood_message (message);
-	}
-	void flood_block (std::shared_ptr<nano::block> block_a)
-	{
-		nano::publish publish (block_a);
-		flood_message (publish);
-	}
-	void flood_block_batch (std::deque<std::shared_ptr<nano::block>>, unsigned = broadcast_interval_ms);
-	void merge_peers (std::array<nano::endpoint, 8> const &);
-	void merge_peer (nano::endpoint const &);
-	void send_keepalive (std::shared_ptr<nano::transport::channel>);
-	void send_keepalive_self (std::shared_ptr<nano::transport::channel>);
-	void send_node_id_handshake (std::shared_ptr<nano::transport::channel>, boost::optional<nano::uint256_union> const & query, boost::optional<nano::uint256_union> const & respond_to);
-	void broadcast_confirm_req (std::shared_ptr<nano::block>);
-	void broadcast_confirm_req_base (std::shared_ptr<nano::block>, std::shared_ptr<std::vector<std::shared_ptr<nano::transport::channel>>>, unsigned, bool = false);
-	void broadcast_confirm_req_batch (std::unordered_map<std::shared_ptr<nano::transport::channel>, std::vector<std::pair<nano::block_hash, nano::block_hash>>>, unsigned = broadcast_interval_ms, bool = false);
-	void broadcast_confirm_req_batch (std::deque<std::pair<std::shared_ptr<nano::block>, std::shared_ptr<std::vector<std::shared_ptr<nano::transport::channel>>>>>, unsigned = broadcast_interval_ms);
-	void confirm_hashes (nano::transaction const &, std::shared_ptr<nano::transport::channel>, std::vector<nano::block_hash>);
-	bool send_votes_cache (std::shared_ptr<nano::transport::channel>, nano::block_hash const &);
-	std::shared_ptr<nano::transport::channel> find_node_id (nano::account const &);
-	std::shared_ptr<nano::transport::channel> find_channel (nano::endpoint const &);
-	bool not_a_peer (nano::endpoint const &, bool);
-	// Should we reach out to this endpoint with a keepalive message
-	bool reachout (nano::endpoint const &, bool = false);
-	std::deque<std::shared_ptr<nano::transport::channel>> list (size_t);
-	// A list of random peers sized for the configured rebroadcast fanout
-	std::deque<std::shared_ptr<nano::transport::channel>> list_fanout ();
-	void random_fill (std::array<nano::endpoint, 8> &) const;
-	std::unordered_set<std::shared_ptr<nano::transport::channel>> random_set (size_t) const;
-	// Get the next peer for attempting a tcp bootstrap connection
-	nano::tcp_endpoint bootstrap_peer ();
-	// Response channels
-	void add_response_channels (nano::tcp_endpoint const &, std::vector<nano::tcp_endpoint>);
-	std::shared_ptr<nano::transport::channel> search_response_channel (nano::tcp_endpoint const &, nano::account const &);
-	void remove_response_channel (nano::tcp_endpoint const &);
-	size_t response_channels_size ();
-	nano::endpoint endpoint ();
-	void cleanup (std::chrono::steady_clock::time_point const &);
-	void ongoing_cleanup ();
-	size_t size () const;
-	size_t size_sqrt () const;
-	bool empty () const;
-	nano::message_buffer_manager buffer_container;
-	boost::asio::ip::udp::resolver resolver;
-	std::vector<boost::thread> packet_processing_threads;
-	nano::node & node;
-	nano::transport::udp_channels udp_channels;
-	nano::transport::tcp_channels tcp_channels;
-	std::function<void()> disconnect_observer;
-	// Called when a new channel is observed
-	std::function<void(std::shared_ptr<nano::transport::channel>)> channel_observer;
-	static unsigned const broadcast_interval_ms = 10;
-	static size_t const buffer_size = 512;
-	static size_t const confirm_req_hashes_max = 6;
-
-private:
-	std::mutex response_channels_mutex;
-	std::unordered_map<nano::tcp_endpoint, std::vector<nano::tcp_endpoint>> response_channels;
-};
-
 class node_init final
 {
 public:
@@ -365,7 +244,6 @@ public:
 	int store_version ();
 	void receive_confirmed (nano::transaction const &, std::shared_ptr<nano::block>, nano::block_hash const &);
 	void process_confirmed (std::shared_ptr<nano::block>, uint8_t = 0);
-	void process_message (nano::message const &, std::shared_ptr<nano::transport::channel>);
 	void process_active (std::shared_ptr<nano::block>);
 	nano::process_return process (nano::block const &);
 	void keepalive_preconfigured (std::vector<std::string> const &);
