+++ nano/node/transport/transport.hpp
@@ -1,9 +1,18 @@
 #pragma once
 
-#include <boost/asio/buffer.hpp>
-
 #include <nano/node/common.hpp>
+#include <nano/node/socket.hpp>
 #include <nano/node/stats.hpp>
+
+#include <unordered_set>
+
+#include <boost/asio/buffer.hpp>
+#include <boost/multi_index/hashed_index.hpp>
+#include <boost/multi_index/mem_fun.hpp>
+#include <boost/multi_index/member.hpp>
+#include <boost/multi_index/ordered_index.hpp>
+#include <boost/multi_index/random_access_index.hpp>
+#include <boost/multi_index_container.hpp>
 
 namespace nano
 {
