+++ nano/node/transport/udp.cpp
@@ -463,7 +463,7 @@ public:
 			node.network.udp_channels.modify (find_channel, [](std::shared_ptr<nano::transport::channel_udp> channel_a) {
 				channel_a->set_last_packet_received (std::chrono::steady_clock::now ());
 			});
-			node.process_message (message_a, find_channel);
+			node.network.process_message (message_a, find_channel);
 		}
 	}
 	nano::node & node;
